Fix OpenRouter override release blockers
- split reasoning helpers out of openrouter.py to keep the public helper under the repo line cap

- preserve source client credentials for model-only OpenRouter overrides

- add focused regression coverage for the model-only override path
